Rename userTerm to user_term for pep8
